refactor(types): .d.ts 파일로 인터페이스 정리
