+++ src/repositories/comment.repository.ts
@@ -2,7 +2,6 @@ import dataSource from './data-source';
 import { Comment } from '../entities/comment.entity';
 import { CommentDto } from '../entities/dto/comment.dto';
 import { Repository } from 'typeorm';
-import { Pagination } from './feedList.repository';
 
 export class CommentRepository extends Repository<Comment> {
   private static instance: CommentRepository;
