+++ src/repositories/feedSymbol.repository.ts
@@ -2,7 +2,6 @@ import { FeedSymbol } from '../entities/feedSymbol.entity';
 import dataSource from './data-source';
 import { Repository } from 'typeorm';
 import { CustomError } from '../utils/util';
-import { Pagination } from './feedList.repository';
 
 export class FeedSymbolRepository extends Repository<FeedSymbol> {
   private static instance: FeedSymbolRepository;
