+++ src/types/feed.d.ts
@@ -1,0 +1,11 @@
+interface FeedOption {
+  isTemp?: boolean;
+  isAll?: boolean;
+}
+
+interface FeedSymbolCount {
+  feedId: number;
+  symbolId: number;
+  symbol: string;
+  count: number;
+}
