+++ src/types/feedList.d.ts
@@ -1,0 +1,13 @@
+interface FeedListOptions {
+  includeTempFeeds?: boolean;
+  onlyTempFeeds?: boolean;
+}
+interface Pagination {
+  startIndex: number;
+  limit: number;
+}
+
+interface PageCondition {
+  skip: number;
+  take: number;
+}
