+++ src/types/feedSymbol.d.ts
@@ -1,10 +1,10 @@
 import { FeedSymbol } from '../entities/feedSymbol.entity';
 
-export interface AddAndUpdateSymbolToFeedResult {
+interface AddAndUpdateSymbolToFeedResult {
   sort: 'add' | 'update';
   result: any;
 }
-export interface CheckSymbolResult {
+interface CheckSymbolResult {
   checkValue: boolean;
   result: FeedSymbol | null;
 }
