+++ src/types/global.d.ts
@@ -27,13 +27,13 @@ declare global {
     status: number;
   }
 
-  export class MissingDriverError extends Error {
-    name = 'MissingDriverError';
-
-    constructor(driverType: string) {
-      super();
-      Object.setPrototypeOf(this, MissingDriverError.prototype);
-      this.message = `Wrong driver: "${driverType}" given. Supported drivers are: "cordova", "mariadb", "mongodb", "mssql", "mysql", "oracle", "postgres", "sqlite", "sqljs", "react-native".`;
-    }
-  }
+  // export class MissingDriverError extends Error {
+  //   name = 'MissingDriverError';
+  //
+  //   constructor(driverType: string) {
+  //     super();
+  //     Object.setPrototypeOf(this, MissingDriverError.prototype);
+  //     this.message = `Wrong driver: "${driverType}" given. Supported drivers are: "cordova", "mariadb", "mongodb", "mssql", "mysql", "oracle", "postgres", "sqlite", "sqljs", "react-native".`;
+  //   }
+  // }
 }
