+++ src/types/upload.d.ts
@@ -1,0 +1,14 @@
+interface Params {
+  Bucket: string;
+  Key: string;
+}
+
+interface DeleteUploadFiles {
+  uploadFileWithoutFeedId: number[];
+  deleteFileLinksArray: string[];
+}
+
+interface deletableFilesBasket {
+  uploadFileIdsToDelete: number[];
+  fileLinksToDelete: string[];
+}
